requirements.txt
asgi-redis==1.0.0
asgiref==1.0.0
autobahn==0.16.1
channels==0.17.3
constantly==15.1.0
csscompressor==0.9.4
daphne==0.15.0
Django==1.10.3
django-appconf==1.0.2
django-bower==5.2.0
django-compressor==2.1
factory-boy==2.5.2
fake-factory==0.5.3
flake8==2.5.4
incremental==16.10.1
mccabe==0.4.0
msgpack-python==0.4.8
oauthlib==2.0.0
pep8==1.7.0
pyflakes==1.0.0
rcssmin==1.0.6
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
Twisted==16.5.0
txaio==2.5.2
zope.interface==4.3.2
